Changed Baker to be an Interface. AkkaBaker is the first implementation of this.
