+++ baas/src/test/scala/com/ing/baker/baas/http/BAASAPIRunner.scala
@@ -1,7 +1,7 @@
 package com.ing.baker.baas.http
 
 import akka.actor.ActorSystem
-import com.ing.baker.runtime.core.Baker
+import com.ing.baker.runtime.core.{AkkaBaker, Baker}
 
 import scala.concurrent.Await
 import scala.concurrent.duration._
@@ -11,7 +11,7 @@ object BAASAPIRunner extends App {
   val host = "localhost"
   val port = 8081
   val actorSystem = ActorSystem("BAASAPIActorSystem")
-  val baker = new Baker()(actorSystem)
+  val baker = new AkkaBaker()(actorSystem)
 
   //Startup the BAASAPI
   val baasAPI = new BAASAPI(baker, host, port)(actorSystem)
